Fix: EEVEE: Irradiance validity loading
The loading step of the irradiance validity was
left unfinished for some reason. It was loading
the same validity for all 8 corners and copying
the same value for all 4 Z slices.

There is another bug, which is that the default
backface validity of surfels is wrong. This means
that the tests are passing because of this bug
as the validity is always 1 everywhere.

Pull Request: https://projects.blender.org/blender/blender/pulls/147200