Refactor: Library: Be explicit about runtime data copied with ID copying.
Note that this commit keeps the same behavior as before this commit,
except for the newly introduced vector of archived libraries, which is
reset to empty. This behavior may have to be further tweaked still.

Root of the issue was from original commit enabling Library ID copying
(87a4c0d3a80bf90), but the later move to an allocated runtime pointer
using a copy constructor in 87a4c0d3a80bf90 made this behavior even
more confusing/hidden.

In general, what is re-used/copied in runtime data should be explicit,
such that by default new data can be assumed reset to their default
value there in the copy ID.