Geometry Nodes: optimize active gizmo detection
This reduces the constant overhead for the evaluation of the Geometry Nodes
modifier. Currently, it has to detect which gizmos are currently used, because
it has to potentially trigger additional evaluation to make the gizmos work.
Part of that is detecting which inputs are used, because gizmos for inputs that
are not used are hidden. Previously, it would compute whether each input is
used. This patch changes it so that only the inputs that have gizmos are
checked.

This reduces the evaluation time in my test from 2.2ms to 1.5ms.

Pull Request: https://projects.blender.org/blender/blender/pulls/147286