Fix #147274: Compositor modifier strip mask affects other strips
Using a strip mask in a compositor strip modifier causes other strips to
change in color. This is because the code converted the mask image
buffer in place, which affected its other uses. To fix this, convert the
mask out of the place.

Pull Request: https://projects.blender.org/blender/blender/pulls/147446