Fix #147413: Thinner Playhead Line Shadow with Wide Lines
Selecting "Wide Lines" has always made the playhead vertical line
thicker. But with #145081 the line is an odd width, so when at 1X scale
it is three pixels instead of two with default lines, five pixels
instead four with wide line lines. But that one pixel isn't a big
deal, it is that the SHADOW around the line is now calculated based on
the line width when it used to be based on ui scale. This means at
1X scale and wide lines the shadow is two pixels when it used to
be just one. This PR goes back to the old shadow width calculation
which makes the overall obscuring width 2 pixels narrower at 1x wide.

Pull Request: https://projects.blender.org/blender/blender/pulls/147467