I18n: Translate node tree subtypes
In 73fcbaf7c4, the `node_tree_sub_type` enum was given a custom item
function. The items now need to be extracted manually.

In addition, the translation context for this property, "ID", was
removed since it was not useful at a glance, and likely added by
mistake.

Reported by Ye Gui in #43295.