Fix #147551: Cryptomatte not supported in VSE
The Cryptomatte node is not supported in VSE strip modifiers because we
previously limited it to root node trees in scene compositing. This
limitation can be lifted however if the user is using the image mode, so
we remove the poll method and instead give warning when appropriate
based on the mode.

Pull Request: https://projects.blender.org/blender/blender/pulls/147601