Fix #147626: Crash when clicking on the "Rendering" workspace
Make the logic in the image editor drawing robust against configuration
when the image editor has flag "show render result from sequencer scene"
and the sequencer scene being nullptr.

It could happen when user configures image editor is such way and then
removes the sequencer scene.

Or, it could also happen when the image editor has the flag set, which
was exactly the cause of the originally reported issue. So this change
also clears the flag which was expected to be cleared. It could affect
some current files saved in the past day, but since the render operator
sets it based on the way the scene is rendered it is not too bad.

Caused by 76c03744a809365af98f4d8356ebaeb0a5e2e831

Pull Request: https://projects.blender.org/blender/blender/pulls/147630