Fix: VSE: Don't show a menu when adding a new sequencer scene
Pressing the `New` button in the VSE would pop up a menu
showing the different options for creating a scene. This includes
the duplication options (e.g. "Full Copy" creates a full copy of the
active scene). Note that Blender always operates with an active
scene which is the reason for why the operator to create a new
scene has all of these options available.

But this doesn't make sense in the context of the
sequencer when there isn't an existing sequencer scene.

This changes the invoke function of `SCENE_OT_new_sequencer_scene`
to check if a sequencer scene exists, and if it doesn't,
just run the execute function with the default "New" setting
(creating a blank scene) and don't show the popup menu.

Pull Request: https://projects.blender.org/blender/blender/pulls/148216