Fix #147617: Missing versioning for drivers of bone.hide
With the `hide` property now being on the `bPoseChannel`,
drivers and animation that used to target the property on the `Bone`
aren't working anymore.

This PR copies the FCurves from the armature to all objects using
that armature via versioning to fix that.

Pull Request: https://projects.blender.org/blender/blender/pulls/147711