Fix #148429: assert due to legacy mesh pointers
`BLO_write_shared_tag` was added in 7ed85bfe17878d. It expects that the shared
pointer has not been written as normal pointer before. However, in some old
files it was written before, because the deprecated `mesh->dvert` had the same
pointer as the corresponding custom data layer. This and similar pointers on
meshes are not used anymore and can therefore just be cleared in versioning
code.

Pull Request: https://projects.blender.org/blender/blender/pulls/148434