GPU: Shader: Parse Create Info
This is the first step towards getting rid of the
implicit shader source ordering.

Create info files are now treated as source files,
pulling their dependencies.

Note that this patch is not generating the resources macros.
This is for future work. Thus, we still pull the create info
dependencies (aka `typedef_source`). This can pull the
`*_define.hh` file twice.

For the parsing to work, all the macro job for variants of
shaders needed to be unrolled.

Pull Request: https://projects.blender.org/blender/blender/pulls/148399