Fix #148235: Grease Pencil armature deform must modify positions in edit mode
Armature modifier for Grease Pencil curves relied on faulty implementation of
armature deformation in 4.5, which caused a lack of deformation on actual curve
data in edit and weight-paint modes. Edit hints were correctly modified so tools
and selection are working, but the drawing code showed only the original point
positions.

The reason is that the `armature_deform.cc` code before 455a8554e2 was
const_casting the `vert_coord_prev` array internally, with the expectation that
this data was in fact a throwaway buffer. In the case of edit hints being
available (i.e. deformed positions for crazyspace) the code would write to
`vert_coord_prev` instead of `vert_coord`. This works for mesh armature
deformation because the invocation always uses the actual vertex output
positions as the target buffer. But Grease Pencil instead uses edit hint
position arrays as the target data and ignores the actual curve positions when
in edit mode.

This _still_ worked in 4.5 because of the aforementioned write to the `const`
`vert_coord_prev` array. Grease Pencil uses a local buffer for this argument,
which is _shared with the curve positions attribute_. This way, when the
`vert_coord_prev` was overwritten it incidentally also updated the curve
positions, which are not an actual mutable argument of the armature deformation.
In 5.0 this illegal armature deform behavior was removed, but that in turn
exposed the buggy Grease Pencil modifier behavior, since now there is no update
of actual curve data in edit mode.

The solution is to separately write both edit hint data and curve positions in
edit mode. Care must be taken here because the modifier _also_ may convert
Bezier curves, in which case the curve topology changes and the "old position"
array cannot be used any more (no multi-modifier feature in this case).
Technically this is illegal as well because the modifier isn't a true
`DeformOnly` modifier any more, but it should work in most cases.

Pull Request: https://projects.blender.org/blender/blender/pulls/148451