Refactor: Mesh: Access active/default UV maps via functions
To make the transition for #122398 simpler, access the names of active UV
maps with a function. This way fewer places will need to change in one diff.
Also add a function to create a VectorSet of the UV maps on a mesh, and
access UV maps via name instead of index in a couple places (apply the
same change to color attribute in 2 places for consistency).

Pull Request: https://projects.blender.org/blender/blender/pulls/148466