Compositor: Add Sequencer Strip Info node
This patch adds a new Sequencer Strip Info node to the compositor. The
node returns information about the strip that the compositor modifier is
applied on, like the Start and End frames as well as its transform. This
node is only supported in the context of the sequencer and will return
default values otherwise.

Pull Request: https://projects.blender.org/blender/blender/pulls/147855