Fix #148604: stable pointers for data-blocks are order dependent
The issue seems to be cause by stable data-block pointers being created when
they are first used. While this is stable in many cases, it can lead to changing
data-block pointers when the first reference to that data-block is moved around
for whatever reason.

This patches fixes it by precomputing all the stable pointers for all
data-blocks to make them (mostly) order-independent. There is still a slight
dependence for the very unlikely case when the name+lib hash of two data-blocks
happens to be the same.

Pull Request: https://projects.blender.org/blender/blender/pulls/148662